reorganize folder structure to clean things up a bit by adding folders, updating imports, etc.
