+++ src/pages/Quote.tsx
@@ -1,6 +1,6 @@
 import { Stack } from "@mui/material";
-import DataTile from "../components/DataTile";
-import SearchBar from "../components/SearchBar";
+import DataTile from "../components/datatile/DataTile";
+import SearchBar from "../components/search/SearchBar";
 import { useState } from "react";
 
 export default function Quote() {
